# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ponyllm-server",
    "description": "ponyllm gateway admin API. Auth: `Authorization: Bearer <gateway-key>` or `x-api-key: <gateway-key>`. `?token=` query auth is deprecated: it leaks credentials into logs/bookmarks and is disabled in `strict` mode (P2); use header auth. Rotate: `POST /api/admin/auth/rotate` (200 one-time plaintext; 401 invalid key; 404 admin-write disabled; 409 open mode has no credential).",
    "contact": {
      "name": "ponyllm team"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.49"
  },
  "paths": {
...
    "/api/admin/models": {
      "get": {
        "tags": [],
        "operationId": "handle_admin_models",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "handle_admin_create_model",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/models/{name}": {
      "put": {
        "tags": [],
        "operationId": "handle_admin_update_model",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModelPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "operationId": "handle_admin_delete_model",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/admin/providers/{name}/models": {
      "get": {
        "tags": [],
        "operationId": "handle_admin_provider_models",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminPricingMode": {
        "type": "string",
        "enum": [
          "uniform",
          "peak_valley"
        ]
      },
      "AdminPricingPeriod": {
        "type": "object",
        "required": [
          "start_time",
          "end_time",
          "input_price",
          "cached_price",
          "output_price"
        ],
        "properties": {
          "cached_price": {
            "type": "number",
            "format": "double"
          },
          "end_time": {
            "type": "string"
          },
          "include_weekends": {
            "type": "boolean"
          },
          "input_price": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "output_price": {
            "type": "number",
            "format": "double"
          },
          "start_time": {
            "type": "string"
          }
        }
      },
      "AdminRateLimits": {
        "type": "object",
        "description": "Admin wire form of the short-window account rate limits (ADR\n`2026-09-30-unified-quota-metering-governance-kernel`).\n\nEvery field is optional. On a model, an absent field inherits the provider\ndefault; on a provider, an absent field means \"no limit\" for that axis.\n`rpm`/`tpm`/`concurrency` treat `0` identically to `None` (unlimited).",
        "properties": {
          "concurrency": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Max in-flight concurrent requests (`None`/`0` = unlimited).",
            "minimum": 0
          },
          "count_cached": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether cache-hit tokens count toward the TPM budget\n(`None` = `true`)."
          },
          "rpm": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Max requests admitted per sliding window (`None`/`0` = unlimited).",
            "minimum": 0
          },
          "tpm": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Max tokens admitted per sliding window (`None`/`0` = unlimited).",
            "minimum": 0
          },
          "window_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Sliding window size in seconds (`None` = 60).",
            "minimum": 0
          }
        }
      },
...
      "CreateModelPayload": {
        "type": "object",
        "required": [
          "provider",
          "name"
        ],
        "properties": {
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "cached_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "context_window": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "fallbacks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "input_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "input_types": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "max_output": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "output_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "output_types": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "pricing_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminPricingMode"
              }
            ]
          },
          "pricing_periods": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/AdminPricingPeriod"
            }
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Explicit routing preference for this model under this provider\n(larger = tried first among same-named models). `None` defaults to no\npreference (0), preserving pre-priority ordering.",
            "minimum": 0
          },
          "protocol": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": "string"
          },
          "proxy": {
            "type": [
              "string",
              "null"
            ]
          },
          "rate_limits": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminRateLimits",
                "description": "Short-window rate limits for this model (RPM/TPM budget + concurrency\ncap), overriding the provider default field-by-field. `None` = inherit\nprovider default (or unlimited)."
              }
            ]
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "thinking_default": {
            "type": [
              "string",
              "null"
            ]
          },
          "thinking_max": {
            "type": [
              "string",
              "null"
            ]
          },
          "tier": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeout_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional total upstream timeout override (seconds, 60~1800).",
            "minimum": 0
          },
          "top_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        }
      },
...
      "ModelView": {
        "type": "object",
        "required": [
          "provider",
          "name",
          "tier",
          "context_window",
          "thinking_default",
          "thinking_max"
        ],
        "properties": {
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "cached_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "context_window": {
            "type": "string"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "input_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Model-level price overrides (`None` inherits the provider baseline)."
          },
          "input_types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "output_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "output_types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pricing_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminPricingMode"
              }
            ]
          },
          "pricing_periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminPricingPeriod"
            }
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Explicit routing preference for this model under this provider\n(larger = tried first among same-named models; `None` = no preference).",
            "minimum": 0
          },
          "protocol": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": "string"
          },
          "proxy": {
            "type": [
              "string",
              "null"
            ],
            "description": "Per-model outbound proxy override (`Some(url)` = this model dials\nthrough the proxy; `None` = direct / inherits the provider default)."
          },
          "rate_limits": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminRateLimits",
                "description": "Short-window account rate limits for this model (RPM/TPM sliding-window\nbudget + concurrency cap). `None` = no model override; the model\ninherits the provider-level default. See [`AdminRateLimits`]."
              }
            ]
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Model-level default sampling (`None` keeps the request value)."
          },
          "thinking_default": {
            "type": "string"
          },
          "thinking_max": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "top_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        }
      },
...
      "UpdateModelPayload": {
        "type": "object",
        "properties": {
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "cached_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "context_window": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "fallbacks": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "input_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "input_types": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "max_output": {
            "type": [
              "string",
              "null"
            ]
          },
          "output_price": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "output_types": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "pricing_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminPricingMode"
              }
            ]
          },
          "pricing_periods": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/AdminPricingPeriod"
            }
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Explicit routing preference for this model under this provider\n(larger = tried first among same-named models). `None` leaves the\nexisting value untouched on update.",
            "minimum": 0
          },
          "protocol": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "proxy": {
            "type": [
              "string",
              "null"
            ]
          },
          "rate_limits": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminRateLimits",
                "description": "Short-window rate limits for this model (`Some(Some(..))` replaces the\nwhole object; explicit `null` (`Some(None)`) clears the model override\nback to \"inherit provider default\"; absent (`None`) leaves it\nuntouched. All-`None` fields = unlimited for those axes, overriding any\nprovider default)."
              }
            ]
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "thinking_default": {
            "type": [
              "string",
              "null"
            ]
          },
          "thinking_max": {
            "type": [
              "string",
              "null"
            ]
          },
          "tier": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeout_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional total upstream timeout override (seconds, 60~1800).",
            "minimum": 0
          },
          "top_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ponyllm gateway API key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}